+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Changed:
 - dialog to suggest new manfacturer entries now mentions RORs for the provenance url
 - the licenes of the system is now the EUPL-1.2
 - Renamed tsm linking to data linking
+- Related units are now suggested, but all units are selectable for measured quantities
 
 Fixed:
 - SensorML validation for unit name values that contain whitespaces